Add alternative blake2b implementations
