+++ Cargo.toml
@@ -6,6 +6,8 @@ license = "MIT Or Apache-2.0"
 repository = "https://github.com/input-output-hk/ce-ashmaize"
 
 [dependencies]
+blake2 = "0.10"
+blake2b_simd = "1.0.3"
 cryptoxide = "~0.5.1"
 
 [dev-dependencies]
@@ -14,6 +16,9 @@ randomx-rs = "1.3"
 hex = "0.4"
 indicatif = "0.17"
 getrandom = "0.3"
+
+[profile.bench]
+debug = true
 
 [[bench]]
 name = "bench"
